[refactor] userImage: 패키지 위치 올바르게 변경
- 변경 전: domain.salepost.userimage
- 변경 후: domain.userimage
