+++ src/main/java/org/example/ootoutfitoftoday/domain/user/service/command/UserCommandServiceImpl.java
@@ -20,11 +20,11 @@ import org.example.ootoutfitoftoday.domain.user.exception.UserErrorCode;
 import org.example.ootoutfitoftoday.domain.user.exception.UserException;
 import org.example.ootoutfitoftoday.domain.user.repository.UserRepository;
 import org.example.ootoutfitoftoday.domain.user.service.query.UserQueryService;
-import org.example.ootoutfitoftoday.domain.salepost.userimage.entity.UserImage;
-import org.example.ootoutfitoftoday.domain.salepost.userimage.exception.UserImageErrorCode;
-import org.example.ootoutfitoftoday.domain.salepost.userimage.exception.UserImageException;
-import org.example.ootoutfitoftoday.domain.salepost.userimage.service.command.UserImageCommandService;
-import org.example.ootoutfitoftoday.domain.salepost.userimage.service.query.UserImageQueryService;
+import org.example.ootoutfitoftoday.domain.userimage.entity.UserImage;
+import org.example.ootoutfitoftoday.domain.userimage.exception.UserImageErrorCode;
+import org.example.ootoutfitoftoday.domain.userimage.exception.UserImageException;
+import org.example.ootoutfitoftoday.domain.userimage.service.command.UserImageCommandService;
+import org.example.ootoutfitoftoday.domain.userimage.service.query.UserImageQueryService;
 import org.springframework.cache.Cache;
 import org.springframework.cache.CacheManager;
 import org.springframework.cache.annotation.CacheEvict;
