+++ src/main/java/org/example/ootoutfitoftoday/domain/userimage/service/query/UserImageQueryService.java
@@ -1,0 +1,8 @@
+package org.example.ootoutfitoftoday.domain.userimage.service.query;
+
+import org.example.ootoutfitoftoday.domain.userimage.entity.UserImage;
+
+public interface UserImageQueryService {
+
+    UserImage findByIdAndIsDeletedFalse(Long userImageId);
+}
